start updating CLI tool
